初期データをロード

+++ api/models/quiz_result.go
@@ -2,7 +2,7 @@ package models
 
 type QuizResult struct {
   ID          int
-  Weight      int  `json:"weight"`
+  Weight      float64  `json:"weight"`
   QuizID      int
   Quiz        Quiz
   UserID      int
